Make some changes, move console only logs to bot.py
