+++ bot.py
@@ -8,6 +8,7 @@ import datetime
 import glob
 import io
 import os
+import sys
 import threading
 from typing import Self
 
@@ -181,6 +182,55 @@ class TechSupportBot(commands.Bot):
         self.extension_name_list = []
         await self.load_extensions()
 
+    async def on_guild_remove(self: Self, guild: discord.Guild) -> None:
+        """See: https://discordpy.readthedocs.io/en/latest/api.html#discord.on_guild_remove
+
+        Args:
+            guild (discord.Guild): The guild that got removed
+        """
+        await self.logger.send_log(
+            message=f"Left guild with ID {guild.id}",
+            level=LogLevel.INFO,
+            console_only=True,
+        )
+
+    async def on_error(self: Self, event_method: str) -> None:
+        """Catches non-command errors and sends them to the error logger for processing.
+
+        Args:
+            event_method (str): the event method name associated with the error (eg. on_message)
+        """
+        _, exception, _ = sys.exc_info()
+        await self.bot.logger.send_log(
+            message=f"Bot error in {event_method}: {exception}",
+            level=LogLevel.ERROR,
+            exception=exception,
+        )
+
+    async def on_connect(self: Self) -> None:
+        """See: https://discordpy.readthedocs.io/en/latest/api.html#discord.on_connect"""
+        await self.logger.send_log(
+            message="Connected to Discord",
+            level=LogLevel.INFO,
+            console_only=True,
+        )
+
+    async def on_resumed(self: Self) -> None:
+        """See: https://discordpy.readthedocs.io/en/latest/api.html#discord.on_resumed"""
+        await self.logger.send_log(
+            message="Resume connection to discord",
+            level=LogLevel.INFO,
+            console_only=True,
+        )
+
+    async def on_disconnect(self: Self) -> None:
+        """See: https://discordpy.readthedocs.io/en/latest/api.html#discord.on_disconnect"""
+        await self.logger.send_log(
+            message="Disconnected from Discord",
+            level=LogLevel.INFO,
+            console_only=True,
+        )
+
     async def on_guild_join(self: Self, guild: discord.Guild) -> None:
         """Configures a new guild upon joining.
         This registers a new guild config, and starts any loop jobs that are configured
@@ -188,6 +238,12 @@ class TechSupportBot(commands.Bot):
         Args:
             guild (discord.Guild): the guild that was joined
         """
+        await self.logger.send_log(
+            message=f"Joined guild with ID {guild.id}",
+            level=LogLevel.INFO,
+            console_only=True,
+        )
+
         for cog in self.cogs.values():
             if getattr(cog, "COG_TYPE", "").lower() == "loop":
                 try:
